feat: post 관련 class에서 user 관련 기능 수정

+++ src/main/java/com/FitToMe/project/Service/Post/PostDeleteService.java
@@ -1,6 +1,7 @@
 package com.FitToMe.project.Service.Post;
 
 import com.FitToMe.project.Entity.Post;
+import com.FitToMe.project.Entity.User;
 import com.FitToMe.project.Repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,9 +14,13 @@ public class PostDeleteService {
 
     private final PostRepository postRepository;
 
-    public boolean deletePost(Long postId) throws RuntimeException {
+    public boolean deletePost(User user, Long postId) throws RuntimeException {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다"));
+
+        if (!post.getUser().equals(user)) {
+            throw new SecurityException("해당 게시글을 삭제할 수 있는 권한이 없습니다");
+        }
 
         postRepository.deleteById(postId);
         return true;
